requirements.txt
Flask
gunicorn
torch
transformers
pydub
flask-cors
python-dotenv
